+++ src/main/kotlin/racingcar/domain/Moveable.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 interface Moveable {
     fun moving(): Boolean
