+++ src/main/kotlin/racingcar/domain/NamedPosition.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 interface NamedPosition {
     val name: Name
