+++ src/test/kotlin/racingcar/domain/NameTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
