+++ src/test/kotlin/racingcar/view/PrinterTest.kt
@@ -1,7 +1,8 @@
-package racingcar
+package racingcar.view
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.Positions
 import java.io.ByteArrayOutputStream
 
 class PrinterTest {
