+++ src/test/kotlin/racingcar/view/UserInputTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.view
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
